Changed the Container to remember what query each communicator performed so that new values can be checked if they should be sent to a specific communicator or not.
This resolves the issue where a communicator should be getting new data but
doesnt since the new data wasnt part of the last query.

Also split out the container code into a few different files so that its more
readable.
